Fr documentation (#9)
* ai-generated readmes

* adding explicit create/update/delete arguments to prevent accidental changes

* update top-level readme

+++ delete_everything/main.go
@@ -24,6 +24,7 @@ func main() {
 
 	logger.Info("Starting")
 	Scope := flag.String("scope", "", "Comma-separated list of things to delete: projects,applications,groups,presets")
+	Update := flag.Bool("update", false, "Set this to true to actually delete, otherwise only inform what would be deleted")
 
 	httpClient := &http.Client{}
 
@@ -40,6 +41,11 @@ func main() {
 		logger.Fatalf("Scope parameter is required. Run with -h for a listing of options.")
 	}
 
+	if !*Update {
+		logger.Warn("This tool will delete ALL projects/applications/groups/presets in the selected scope. Use with caution, this is irreversible.")
+		logger.Info("Running in informational mode only, no changes will be made (-update flag not set)")
+	}
+
 	logger.Infof("Connected with %v", cx1client.String())
 
 	scopes := strings.Split(strings.ToLower(*Scope), ",")
@@ -47,13 +53,13 @@ func main() {
 	for _, s := range scopes {
 		switch s {
 		case "projects":
-			deleteProjects(cx1client, logger)
+			deleteProjects(cx1client, logger, *Update)
 		case "applications":
-			deleteApplications(cx1client, logger)
+			deleteApplications(cx1client, logger, *Update)
 		case "groups":
-			deleteGroups(cx1client, logger)
+			deleteGroups(cx1client, logger, *Update)
 		case "presets":
-			deletePresets(cx1client, logger)
+			deletePresets(cx1client, logger, *Update)
 		default:
 			logger.Errorf("Unknown scope parameter: %v", s)
 		}
@@ -61,11 +67,15 @@ func main() {
 
 }
 
-func deleteProjects(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger) {
+func deleteProjects(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger, update bool) {
 	if projects, err := cx1client.GetProjects(0); err != nil {
 		logger.Errorf("Failed to get projects: %s", err)
 	} else {
 		for _, project := range projects {
+			if !update {
+				logger.Infof("Would delete project %v", project.String())
+				continue
+			}
 			if err = cx1client.DeleteProject(&project); err != nil {
 				logger.Errorf("Failed to delete project %v: %s", project.String(), err)
 			} else {
@@ -75,11 +85,15 @@ func deleteProjects(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger) {
 	}
 }
 
-func deleteApplications(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger) {
+func deleteApplications(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger, update bool) {
 	if applications, err := cx1client.GetApplications(0); err != nil {
 		logger.Errorf("Failed to get applications: %s", err)
 	} else {
 		for _, application := range applications {
+			if !update {
+				logger.Infof("Would delete application %v", application.String())
+				continue
+			}
 			if err = cx1client.DeleteApplication(&application); err != nil {
 				logger.Errorf("Failed to delete application: %s %v", application.String(), err)
 			} else {
@@ -89,11 +103,15 @@ func deleteApplications(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger)
 	}
 }
 
-func deleteGroups(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger) {
+func deleteGroups(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger, update bool) {
 	if groups, err := cx1client.GetGroups(); err != nil {
 		logger.Errorf("Failed to get groups: %s", err)
 	} else {
 		for _, group := range groups {
+			if !update {
+				logger.Infof("Would delete group %v", group.String())
+				continue
+			}
 			if err = cx1client.DeleteGroup(&group); err != nil {
 				logger.Errorf("Failed to delete group %v: %s", group.String(), err)
 			} else {
@@ -103,7 +121,7 @@ func deleteGroups(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger) {
 	}
 }
 
-func deletePresets(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger) {
+func deletePresets(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger, update bool) {
 	if count, err := cx1client.GetPresetCount(); err != nil {
 		logger.Errorf("Failed to get preset count: %s", err)
 	} else {
@@ -111,6 +129,10 @@ func deletePresets(cx1client *Cx1ClientGo.Cx1Client, logger *logrus.Logger) {
 			logger.Errorf("Failed to get presets: %s", err)
 		} else {
 			for _, preset := range presets {
+				if !update {
+					logger.Infof("Would delete preset %v", preset.String())
+					continue
+				}
 				if err = cx1client.DeletePreset(&preset); err != nil {
 					logger.Errorf("Failed to delete preset %v: %s", preset.String(), err)
 				} else {
